Make add_output_file with validate_class available to oceanmodelstep and oceaniostep
